More docker tweaks but still not working
I so far have been unable to get the rake db tasks to run when upping
the container. The goal here is to automate db creation so the dev
doesn't need to run any commands after `compose up`. Rake doesn't like
that the postgres instance is "remote". This works fine in an app at
work so I know it's possible - I'm just not sure what I'm missing to
make it work here.